src/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
apache-beam==2.46.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.0
blessed==1.20.0
blis==0.7.9
cachetools==4.2.4
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.0
confection==0.0.4
contourpy==1.0.7
crcmod==1.7
croniter==1.3.8
cycler==0.11.0
cymem==2.0.7
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.3.0
dill==0.3.1.1
dnspython==2.3.0
docopt==0.6.2
email-validator==1.3.1
executing==1.2.0
fastapi==0.88.0
fastavro==1.7.3
fasteners==0.18
filelock==3.10.2
flatbuffers==23.3.3
fonttools==4.39.2
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.5.0/fr_core_news_sm-3.5.0-py3-none-any.whl
frozenlist==1.3.3
fsspec==2023.3.0
gast==0.4.0
google-api-core==2.11.0
google-api-python-client==1.12.11
google-apitools==0.5.31
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.7.0
google-cloud-bigquery-storage==2.16.2
google-cloud-bigtable==1.7.3
google-cloud-core==2.3.2
google-cloud-datastore==1.15.5
google-cloud-dlp==3.12.0
google-cloud-language==1.3.2
google-cloud-pubsub==2.15.1
google-cloud-pubsublite==1.8.0
google-cloud-recommendations-ai==0.7.1
google-cloud-spanner==3.28.0
google-cloud-videointelligence==1.16.3
google-cloud-vision==3.4.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.3
grpcio-status==1.48.2
h11==0.14.0
h5py==3.8.0
hdfs==2.7.0
httpcore==0.16.3
httplib2==0.21.0
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.13.3
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
inquirer==3.1.3
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
langcodes==3.3.0
libclang==15.0.6.1
lightning==2.0.0
lightning-cloud==0.5.32
lightning-utilities==0.8.0
lit==16.0.0
llvmlite==0.39.1
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
networkx==3.0
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauth2client==4.1.3
oauthlib==3.2.2
objsize==0.6.1
omegaconf==2.3.0
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.8.8
overrides==6.5.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
preshed==3.0.8
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
pydot==1.4.2
Pygments==2.14.0
PyJWT==2.6.0
pymongo==3.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.0
pytz==2022.7.1
PyYAML==6.0
readchar==4.0.5
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.3.3
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
shap==0.41.0
six==1.16.0
slicer==0.0.7
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.6
stack-data==0.6.2
starlette==0.22.0
starsessions==1.3.0
sympy==1.11.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.1
tensorflow-estimator==2.11.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.12.0
tensorflow-serving-api==2.11.0
tensorflow-text==2.11.0
tensorflow-transform==1.12.0
termcolor==2.2.0
tfx-bsl==1.12.0
thinc==8.1.9
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==2.0.0
torchdata==0.6.0
torchmetrics==0.11.4
torchtext==0.15.1
torchvision==0.15.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.3
triton==2.0.0
typer==0.7.0
typing_extensions==4.5.0
ujson==5.7.0
Unidecode==1.3.6
uritemplate==3.0.1
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
wasabi==1.1.1
watchfiles==0.18.1
wcwidth==0.2.6
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.8.2
zipp==3.15.0
zstandard==0.20.0
